+++ src/components/App.js
@@ -1,16 +1,8 @@
 import React, { Component } from 'react';
 // import { connect } from 'react-redux';
-import Header from './Header';
+import HeaderContainer from '../containers/HeaderContainer'
 
 class App extends Component {
-
-  // getCurrentLocation(){
-  //   navigator.geolocation.getCurrentPosition((data) => {
-  //      this.props.fetchLocation(data).then(()=>{
-  //        return this.getWeather()
-  //      })
-  //   })
-  // }
 
   getWeather(weatherURL){
     fetch(weatherURL)
@@ -27,8 +19,6 @@ class App extends Component {
     navigator.geolocation.getCurrentPosition((position) => {
       const weatherURL = `http://api.wunderground.com/api/6fc8de6a49f48b06/geolookup/forecast/hourly/forecast10day/conditions/q/${position.coords.latitude},${position.coords.longitude}.json`
       this.getWeather(weatherURL)
-// json.current_observation.temp_f
-// json.current_observation.overcast
       })
     }
 
@@ -36,9 +26,8 @@ class App extends Component {
     render(){
       return (
         <div>
-        HELLO
-      {this.props.weather ? <div>{this.props.weather.temp} {this.props.weather.currently}</div> : <div>LOADING MOTHAFUCKA</div> }
-    </div>
+          <HeaderContainer />
+      </div>
       )
     }
   }
