--- src/containers/Header.js
@@ -1,11 +1,0 @@
-// import { connect } from 'react-redux';
-// import {receiveLocation} from '../actions'
-// import Header from '../components/Header'
-//
-// const mapStateToProps = state => {
-//   return {
-//     location: state.location,
-//   };
-// }
-//
-// export default connect(mapStateToProps, null)(Header);
