+++ src/containers/HeaderContainer.js
@@ -1,0 +1,11 @@
+import { connect } from 'react-redux';
+import {currentWeather} from '../actions'
+import Header from '../components/Header'
+
+const mapStateToProps = state => {
+  return {
+    weather: state.currentWeatherReducer,
+  }
+}
+
+export default connect(mapStateToProps, null)(Header);
