+++ src/routes.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Router, Route } from 'react-router';
 
-import App from './containers/App';
+import AppContainer from './containers/AppContainer';
 
 const Routes = (props) => (
   <Router {...props}>
-    <Route path="/" component={App} />
+    <Route path="/" component={AppContainer} />
   </Router>
 );
